% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About.additionalRoleInfo}
\alias{About.additionalRoleInfo}
\title{About.additionalRoleInfo Object}
\usage{
About.additionalRoleInfo(About.additionalRoleInfo.roleSets = NULL)
}
\arguments{
\item{About.additionalRoleInfo.roleSets}{The \link{About.additionalRoleInfo.roleSets} object or list of objects}
}
\value{
About.additionalRoleInfo object
}
\description{
About.additionalRoleInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other About functions: \code{\link{About.additionalRoleInfo.roleSets}},
  \code{\link{About.exportFormats}},
  \code{\link{About.features}},
  \code{\link{About.importFormats}},
  \code{\link{About.maxUploadSizes}},
  \code{\link{About.quotaBytesByService}},
  \code{\link{About}}
}

