% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About.additionalRoleInfo.roleSets}
\alias{About.additionalRoleInfo.roleSets}
\title{About.additionalRoleInfo.roleSets Object}
\usage{
About.additionalRoleInfo.roleSets()
}
\value{
About.additionalRoleInfo.roleSets object
}
\description{
About.additionalRoleInfo.roleSets Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other About functions: \code{\link{About.additionalRoleInfo}},
  \code{\link{About.exportFormats}},
  \code{\link{About.features}},
  \code{\link{About.importFormats}},
  \code{\link{About.maxUploadSizes}},
  \code{\link{About.quotaBytesByService}},
  \code{\link{About}}
}
