% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About.exportFormats}
\alias{About.exportFormats}
\title{About.exportFormats Object}
\usage{
About.exportFormats()
}
\value{
About.exportFormats object
}
\description{
About.exportFormats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other About functions: \code{\link{About.additionalRoleInfo.roleSets}},
  \code{\link{About.additionalRoleInfo}},
  \code{\link{About.features}},
  \code{\link{About.importFormats}},
  \code{\link{About.maxUploadSizes}},
  \code{\link{About.quotaBytesByService}},
  \code{\link{About}}
}

