% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About.features}
\alias{About.features}
\title{About.features Object}
\usage{
About.features()
}
\value{
About.features object
}
\description{
About.features Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other About functions: \code{\link{About.additionalRoleInfo.roleSets}},
  \code{\link{About.additionalRoleInfo}},
  \code{\link{About.exportFormats}},
  \code{\link{About.importFormats}},
  \code{\link{About.maxUploadSizes}},
  \code{\link{About.quotaBytesByService}},
  \code{\link{About}}
}

