% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{about.get}
\alias{about.get}
\title{Gets the information about the current user along with Drive API settings}
\usage{
about.get(includeSubscribed = NULL, maxChangeIdCount = NULL,
  startChangeId = NULL)
}
\arguments{
\item{includeSubscribed}{When calculating the number of remaining change IDs, whether to include public files the user has opened and shared files}

\item{maxChangeIdCount}{Maximum number of remaining change IDs to count}

\item{startChangeId}{Change ID to start counting from when calculating number of remaining change IDs}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.metadata.readonly, https://www.googleapis.com/auth/drive.photos.readonly, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}

