% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About.maxUploadSizes}
\alias{About.maxUploadSizes}
\title{About.maxUploadSizes Object}
\usage{
About.maxUploadSizes()
}
\value{
About.maxUploadSizes object
}
\description{
About.maxUploadSizes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other About functions: \code{\link{About.additionalRoleInfo.roleSets}},
  \code{\link{About.additionalRoleInfo}},
  \code{\link{About.exportFormats}},
  \code{\link{About.features}},
  \code{\link{About.importFormats}},
  \code{\link{About.quotaBytesByService}},
  \code{\link{About}}
}

