% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About}
\alias{About}
\title{Drive API Objects 
Manages files in Drive including uploading, downloading, searching, detecting changes, and updating sharing permissions.}
\usage{
About(About.additionalRoleInfo = NULL,
  About.additionalRoleInfo.roleSets = NULL, About.exportFormats = NULL,
  About.features = NULL, About.importFormats = NULL,
  About.maxUploadSizes = NULL, About.quotaBytesByService = NULL,
  additionalRoleInfo = NULL, domainSharingPolicy = NULL, etag = NULL,
  exportFormats = NULL, features = NULL, folderColorPalette = NULL,
  importFormats = NULL, isCurrentAppInstalled = NULL, languageCode = NULL,
  largestChangeId = NULL, maxUploadSizes = NULL, name = NULL,
  permissionId = NULL, quotaBytesByService = NULL, quotaBytesTotal = NULL,
  quotaBytesUsed = NULL, quotaBytesUsedAggregate = NULL,
  quotaBytesUsedInTrash = NULL, quotaType = NULL,
  remainingChangeIds = NULL, rootFolderId = NULL, selfLink = NULL,
  user = NULL)
}
\arguments{
\item{About.additionalRoleInfo}{The \link{About.additionalRoleInfo} object or list of objects}

\item{About.additionalRoleInfo.roleSets}{The \link{About.additionalRoleInfo.roleSets} object or list of objects}

\item{About.exportFormats}{The \link{About.exportFormats} object or list of objects}

\item{About.features}{The \link{About.features} object or list of objects}

\item{About.importFormats}{The \link{About.importFormats} object or list of objects}

\item{About.maxUploadSizes}{The \link{About.maxUploadSizes} object or list of objects}

\item{About.quotaBytesByService}{The \link{About.quotaBytesByService} object or list of objects}

\item{additionalRoleInfo}{Information about supported additional roles per file type}

\item{domainSharingPolicy}{The domain sharing policy for the current user}

\item{etag}{The ETag of the item}

\item{exportFormats}{The allowable export formats}

\item{features}{List of additional features enabled on this account}

\item{folderColorPalette}{The palette of allowable folder colors as RGB hex strings}

\item{importFormats}{The allowable import formats}

\item{isCurrentAppInstalled}{A boolean indicating whether the authenticated app is installed by the authenticated user}

\item{languageCode}{The user's language or locale code, as defined by BCP 47, with some extensions from Unicode's LDML format (http://www}

\item{largestChangeId}{The largest change id}

\item{maxUploadSizes}{List of max upload sizes for each file type}

\item{name}{The name of the current user}

\item{permissionId}{The current user's ID as visible in the permissions collection}

\item{quotaBytesByService}{The amount of storage quota used by different Google services}

\item{quotaBytesTotal}{The total number of quota bytes}

\item{quotaBytesUsed}{The number of quota bytes used by Google Drive}

\item{quotaBytesUsedAggregate}{The number of quota bytes used by all Google apps (Drive, Picasa, etc}

\item{quotaBytesUsedInTrash}{The number of quota bytes used by trashed items}

\item{quotaType}{The type of the user's storage quota}

\item{remainingChangeIds}{The number of remaining change ids, limited to no more than 2500}

\item{rootFolderId}{The id of the root folder}

\item{selfLink}{A link back to this item}

\item{user}{The authenticated user}
}
\value{
About object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:11:51
filename: /Users/mark/dev/R/autoGoogleAPI/googledrivev2.auto/R/drive_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
About Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An item with user information and settings.
}
\seealso{
Other About functions: \code{\link{About.additionalRoleInfo.roleSets}},
  \code{\link{About.additionalRoleInfo}},
  \code{\link{About.exportFormats}},
  \code{\link{About.features}},
  \code{\link{About.importFormats}},
  \code{\link{About.maxUploadSizes}},
  \code{\link{About.quotaBytesByService}}
}

