% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{App.icons}
\alias{App.icons}
\title{App.icons Object}
\usage{
App.icons()
}
\value{
App.icons object
}
\description{
App.icons Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other App functions: \code{\link{App}}
}

