% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{App}
\alias{App}
\title{App Object}
\usage{
App(App.icons = NULL, authorized = NULL, createInFolderTemplate = NULL,
  createUrl = NULL, hasDriveWideScope = NULL, icons = NULL, id = NULL,
  installed = NULL, longDescription = NULL, name = NULL,
  objectType = NULL, openUrlTemplate = NULL, primaryFileExtensions = NULL,
  primaryMimeTypes = NULL, productId = NULL, productUrl = NULL,
  secondaryFileExtensions = NULL, secondaryMimeTypes = NULL,
  shortDescription = NULL, supportsCreate = NULL, supportsImport = NULL,
  supportsMultiOpen = NULL, supportsOfflineCreate = NULL,
  useByDefault = NULL)
}
\arguments{
\item{App.icons}{The \link{App.icons} object or list of objects}

\item{authorized}{Whether the app is authorized to access data on the user's Drive}

\item{createInFolderTemplate}{The template url to create a new file with this app in a given folder}

\item{createUrl}{The url to create a new file with this app}

\item{hasDriveWideScope}{Whether the app has drive-wide scope}

\item{icons}{The various icons for the app}

\item{id}{The ID of the app}

\item{installed}{Whether the app is installed}

\item{longDescription}{A long description of the app}

\item{name}{The name of the app}

\item{objectType}{The type of object this app creates (e}

\item{openUrlTemplate}{The template url for opening files with this app}

\item{primaryFileExtensions}{The list of primary file extensions}

\item{primaryMimeTypes}{The list of primary mime types}

\item{productId}{The ID of the product listing for this app}

\item{productUrl}{A link to the product listing for this app}

\item{secondaryFileExtensions}{The list of secondary file extensions}

\item{secondaryMimeTypes}{The list of secondary mime types}

\item{shortDescription}{A short description of the app}

\item{supportsCreate}{Whether this app supports creating new objects}

\item{supportsImport}{Whether this app supports importing Google Docs}

\item{supportsMultiOpen}{Whether this app supports opening more than one file}

\item{supportsOfflineCreate}{Whether this app supports creating new files when offline}

\item{useByDefault}{Whether the app is selected as the default handler for the types it supports}
}
\value{
App object
}
\description{
App Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The apps resource provides a list of the apps that a user has installed, with information about each app's supported MIME types, file extensions, and other details.
}
\seealso{
Other App functions: \code{\link{App.icons}}
}

