% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{AppList}
\alias{AppList}
\title{AppList Object}
\usage{
AppList(defaultAppIds = NULL, etag = NULL, items = NULL,
  selfLink = NULL)
}
\arguments{
\item{defaultAppIds}{List of app IDs that the user has specified to use by default}

\item{etag}{The ETag of the list}

\item{items}{The list of apps}

\item{selfLink}{A link back to this list}
}
\value{
AppList object
}
\description{
AppList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of third-party applications which the user has installed or given access to Google Drive.
}
