% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{ChangeList}
\alias{ChangeList}
\title{ChangeList Object}
\usage{
ChangeList(etag = NULL, items = NULL, largestChangeId = NULL,
  nextLink = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{etag}{The ETag of the list}

\item{items}{The actual list of changes}

\item{largestChangeId}{The current largest change ID}

\item{nextLink}{A link to the next page of changes}

\item{nextPageToken}{The page token for the next page of changes}

\item{selfLink}{A link back to this list}
}
\value{
ChangeList object
}
\description{
ChangeList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of changes for a user.
}

