% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{changes.watch}
\alias{changes.watch}
\title{Subscribe to changes for a user.}
\usage{
changes.watch(Channel, includeDeleted = NULL, includeSubscribed = NULL,
  maxResults = NULL, pageToken = NULL, spaces = NULL,
  startChangeId = NULL)
}
\arguments{
\item{Channel}{The \link{Channel} object to pass to this method}

\item{includeDeleted}{Whether to include deleted items}

\item{includeSubscribed}{Whether to include public files the user has opened and shared files}

\item{maxResults}{Maximum number of changes to return}

\item{pageToken}{Page token for changes}

\item{spaces}{A comma-separated list of spaces to query}

\item{startChangeId}{Change ID to start listing changes from}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.apps.readonly
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.apps.readonly, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.metadata.readonly, https://www.googleapis.com/auth/drive.photos.readonly, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other Channel functions: \code{\link{Channel.params}},
  \code{\link{Channel}}, \code{\link{channels.stop}},
  \code{\link{files.watch}}
}

