% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Channel.params}
\alias{Channel.params}
\title{Channel.params Object}
\usage{
Channel.params()
}
\value{
Channel.params object
}
\description{
Channel.params Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional parameters controlling delivery channel behavior. Optional.
}
\seealso{
Other Channel functions: \code{\link{Channel}},
  \code{\link{changes.watch}}, \code{\link{channels.stop}},
  \code{\link{files.watch}}
}
