% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{ChildList}
\alias{ChildList}
\title{ChildList Object}
\usage{
ChildList(etag = NULL, items = NULL, nextLink = NULL,
  nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{etag}{The ETag of the list}

\item{items}{The actual list of children}

\item{nextLink}{A link to the next page of children}

\item{nextPageToken}{The page token for the next page of children}

\item{selfLink}{A link back to this list}
}
\value{
ChildList object
}
\description{
ChildList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of children of a file.
}

