% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{ChildReference}
\alias{ChildReference}
\title{ChildReference Object}
\usage{
ChildReference(childLink = NULL, id = NULL, selfLink = NULL)
}
\arguments{
\item{childLink}{A link to the child}

\item{id}{The ID of the child}

\item{selfLink}{A link back to this reference}
}
\value{
ChildReference object
}
\description{
ChildReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference to a folder's child.
}
\seealso{
Other ChildReference functions: \code{\link{children.insert}}
}
