% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{children.insert}
\alias{children.insert}
\title{Inserts a file into a folder.}
\usage{
children.insert(ChildReference, folderId)
}
\arguments{
\item{ChildReference}{The \link{ChildReference} object to pass to this method}

\item{folderId}{The ID of the folder}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other ChildReference functions: \code{\link{ChildReference}}
}

