% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{children.list}
\alias{children.list}
\title{Lists a folder's children.}
\usage{
children.list(folderId, maxResults = NULL, orderBy = NULL,
  pageToken = NULL, q = NULL)
}
\arguments{
\item{folderId}{The ID of the folder}

\item{maxResults}{Maximum number of children to return}

\item{orderBy}{A comma-separated list of sort keys}

\item{pageToken}{Page token for children}

\item{q}{Query string for searching children}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.metadata.readonly, https://www.googleapis.com/auth/drive.photos.readonly, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}

