% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Comment.context}
\alias{Comment.context}
\title{Comment.context Object}
\usage{
Comment.context(type = NULL, value = NULL)
}
\arguments{
\item{type}{The MIME type of the context snippet}

\item{value}{Data representation of the segment of the file being commented on}
}
\value{
Comment.context object
}
\description{
Comment.context Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The context of the file which is being commented on.
}
\seealso{
Other Comment functions: \code{\link{Comment}},
  \code{\link{comments.insert}},
  \code{\link{comments.patch}},
  \code{\link{comments.update}}
}
