% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Comment}
\alias{Comment}
\title{Comment Object}
\usage{
Comment(Comment.context = NULL, anchor = NULL, author = NULL,
  commentId = NULL, content = NULL, context = NULL, createdDate = NULL,
  deleted = NULL, fileId = NULL, fileTitle = NULL, htmlContent = NULL,
  modifiedDate = NULL, replies = NULL, selfLink = NULL, status = NULL)
}
\arguments{
\item{Comment.context}{The \link{Comment.context} object or list of objects}

\item{anchor}{A region of the document represented as a JSON string}

\item{author}{The user who wrote this comment}

\item{commentId}{The ID of the comment}

\item{content}{The plain text content used to create this comment}

\item{context}{The context of the file which is being commented on}

\item{createdDate}{The date when this comment was first created}

\item{deleted}{Whether this comment has been deleted}

\item{fileId}{The file which this comment is addressing}

\item{fileTitle}{The title of the file which this comment is addressing}

\item{htmlContent}{HTML formatted content for this comment}

\item{modifiedDate}{The date when this comment or any of its replies were last modified}

\item{replies}{Replies to this post}

\item{selfLink}{A link back to this comment}

\item{status}{The status of this comment}
}
\value{
Comment object
}
\description{
Comment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A comment on a file in Google Drive.
}
\seealso{
Other Comment functions: \code{\link{Comment.context}},
  \code{\link{comments.insert}},
  \code{\link{comments.patch}},
  \code{\link{comments.update}}
}
