% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{CommentList}
\alias{CommentList}
\title{CommentList Object}
\usage{
CommentList(items = NULL, nextLink = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{items}{List of comments}

\item{nextLink}{A link to the next page of comments}

\item{nextPageToken}{The token to use to request the next page of results}

\item{selfLink}{A link back to this list}
}
\value{
CommentList object
}
\description{
CommentList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of comments on a file in Google Drive.
}

