% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{comments.list}
\alias{comments.list}
\title{Lists a file's comments.}
\usage{
comments.list(fileId, includeDeleted = NULL, maxResults = NULL,
  pageToken = NULL, updatedMin = NULL)
}
\arguments{
\item{fileId}{The ID of the file}

\item{includeDeleted}{If set, all comments and replies, including deleted comments and replies (with content stripped) will be returned}

\item{maxResults}{The maximum number of discussions to include in the response, used for paging}

\item{pageToken}{The continuation token, used to page through large result sets}

\item{updatedMin}{Only discussions that were updated after this timestamp will be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}

