% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.capabilities}
\alias{File.capabilities}
\title{File.capabilities Object}
\usage{
File.capabilities(canAddChildren = NULL, canComment = NULL,
  canCopy = NULL, canDelete = NULL, canDownload = NULL, canEdit = NULL,
  canListChildren = NULL, canMoveItemIntoTeamDrive = NULL,
  canMoveTeamDriveItem = NULL, canReadRevisions = NULL,
  canReadTeamDrive = NULL, canRemoveChildren = NULL, canRename = NULL,
  canShare = NULL, canTrash = NULL, canUntrash = NULL)
}
\arguments{
\item{canAddChildren}{Whether the current user can add children to this folder}

\item{canComment}{Whether the current user can comment on the file}

\item{canCopy}{Whether the file can be copied by the current user}

\item{canDelete}{Whether the file can be deleted by the current user}

\item{canDownload}{Whether the file can be downloaded by the current user}

\item{canEdit}{Whether the file can be edited by the current user}

\item{canListChildren}{Whether the current user can list the children of this folder}

\item{canMoveItemIntoTeamDrive}{Whether the current user can move this item into a Team Drive}

\item{canMoveTeamDriveItem}{Whether the current user can move this Team Drive item by changing its parent}

\item{canReadRevisions}{Whether the current user has read access to the Revisions resource of the file}

\item{canReadTeamDrive}{Whether the current user has read access to the Team Drive to which this file belongs}

\item{canRemoveChildren}{Whether the current user can remove children from this folder}

\item{canRename}{Whether the file can be renamed by the current user}

\item{canShare}{Whether the file's sharing settings can be modified by the current user}

\item{canTrash}{Whether the file can be trashed by the current user}

\item{canUntrash}{Whether the file can be restored from the trash by the current user}
}
\value{
File.capabilities object
}
\description{
File.capabilities Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Capabilities the current user has on the file. Each capability corresponds to a fine-grained action that a user may take.
}
\seealso{
Other File functions: \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.insert}}, \code{\link{files.patch}},
  \code{\link{files.update}}
}
