% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.exportLinks}
\alias{File.exportLinks}
\title{File.exportLinks Object}
\usage{
File.exportLinks()
}
\value{
File.exportLinks object
}
\description{
File.exportLinks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Links for exporting Google Docs to specific formats.
}
\seealso{
Other File functions: \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.insert}}, \code{\link{files.patch}},
  \code{\link{files.update}}
}

