% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.imageMediaMetadata.location}
\alias{File.imageMediaMetadata.location}
\title{File.imageMediaMetadata.location Object}
\usage{
File.imageMediaMetadata.location(altitude = NULL, latitude = NULL,
  longitude = NULL)
}
\arguments{
\item{altitude}{The altitude stored in the image}

\item{latitude}{The latitude stored in the image}

\item{longitude}{The longitude stored in the image}
}
\value{
File.imageMediaMetadata.location object
}
\description{
File.imageMediaMetadata.location Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Geographic location information stored in the image.
}
\seealso{
Other File functions: \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.insert}}, \code{\link{files.patch}},
  \code{\link{files.update}}
}

