% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.imageMediaMetadata}
\alias{File.imageMediaMetadata}
\title{File.imageMediaMetadata Object}
\usage{
File.imageMediaMetadata(File.imageMediaMetadata.location = NULL,
  aperture = NULL, cameraMake = NULL, cameraModel = NULL,
  colorSpace = NULL, date = NULL, exposureBias = NULL,
  exposureMode = NULL, exposureTime = NULL, flashUsed = NULL,
  focalLength = NULL, height = NULL, isoSpeed = NULL, lens = NULL,
  location = NULL, maxApertureValue = NULL, meteringMode = NULL,
  rotation = NULL, sensor = NULL, subjectDistance = NULL,
  whiteBalance = NULL, width = NULL)
}
\arguments{
\item{File.imageMediaMetadata.location}{The \link{File.imageMediaMetadata.location} object or list of objects}

\item{aperture}{The aperture used to create the photo (f-number)}

\item{cameraMake}{The make of the camera used to create the photo}

\item{cameraModel}{The model of the camera used to create the photo}

\item{colorSpace}{The color space of the photo}

\item{date}{The date and time the photo was taken (EXIF format timestamp)}

\item{exposureBias}{The exposure bias of the photo (APEX value)}

\item{exposureMode}{The exposure mode used to create the photo}

\item{exposureTime}{The length of the exposure, in seconds}

\item{flashUsed}{Whether a flash was used to create the photo}

\item{focalLength}{The focal length used to create the photo, in millimeters}

\item{height}{The height of the image in pixels}

\item{isoSpeed}{The ISO speed used to create the photo}

\item{lens}{The lens used to create the photo}

\item{location}{Geographic location information stored in the image}

\item{maxApertureValue}{The smallest f-number of the lens at the focal length used to create the photo (APEX value)}

\item{meteringMode}{The metering mode used to create the photo}

\item{rotation}{The rotation in clockwise degrees from the image's original orientation}

\item{sensor}{The type of sensor used to create the photo}

\item{subjectDistance}{The distance to the subject of the photo, in meters}

\item{whiteBalance}{The white balance mode used to create the photo}

\item{width}{The width of the image in pixels}
}
\value{
File.imageMediaMetadata object
}
\description{
File.imageMediaMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata about image media. This will only be present for image types, and its contents will depend on what can be parsed from the image content.
}
\seealso{
Other File functions: \code{\link{File.capabilities}},
  \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.insert}}, \code{\link{files.patch}},
  \code{\link{files.update}}
}
