% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.indexableText}
\alias{File.indexableText}
\title{File.indexableText Object}
\usage{
File.indexableText(text = NULL)
}
\arguments{
\item{text}{The text to be indexed for this file}
}
\value{
File.indexableText object
}
\description{
File.indexableText Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indexable text attributes for the file (can only be written)
}
\seealso{
Other File functions: \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.insert}}, \code{\link{files.patch}},
  \code{\link{files.update}}
}

