% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.labels}
\alias{File.labels}
\title{File.labels Object}
\usage{
File.labels(hidden = NULL, modified = NULL, restricted = NULL,
  starred = NULL, trashed = NULL, viewed = NULL)
}
\arguments{
\item{hidden}{Deprecated}

\item{modified}{Whether the file has been modified by this user}

\item{restricted}{Whether viewers and commenters are prevented from downloading, printing, and copying this file}

\item{starred}{Whether this file is starred by the user}

\item{trashed}{Whether this file has been trashed}

\item{viewed}{Whether this file has been viewed by this user}
}
\value{
File.labels object
}
\description{
File.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group of labels for the file.
}
\seealso{
Other File functions: \code{\link{File.capabilities}},
  \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.insert}}, \code{\link{files.patch}},
  \code{\link{files.update}}
}
