% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.openWithLinks}
\alias{File.openWithLinks}
\title{File.openWithLinks Object}
\usage{
File.openWithLinks()
}
\value{
File.openWithLinks object
}
\description{
File.openWithLinks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of the id of each of the user's apps to a link to open this file with that app. Only populated when the drive.apps.readonly scope is used.
}
\seealso{
Other File functions: \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}}, \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.insert}}, \code{\link{files.patch}},
  \code{\link{files.update}}
}

