% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File}
\alias{File}
\title{File Object}
\usage{
File(File.capabilities = NULL, File.exportLinks = NULL,
  File.imageMediaMetadata = NULL, File.imageMediaMetadata.location = NULL,
  File.indexableText = NULL, File.labels = NULL,
  File.openWithLinks = NULL, File.thumbnail = NULL,
  File.videoMediaMetadata = NULL, alternateLink = NULL,
  appDataContents = NULL, canComment = NULL, canReadRevisions = NULL,
  capabilities = NULL, copyable = NULL, createdDate = NULL,
  defaultOpenWithLink = NULL, description = NULL, downloadUrl = NULL,
  editable = NULL, embedLink = NULL, etag = NULL,
  explicitlyTrashed = NULL, exportLinks = NULL, fileExtension = NULL,
  fileSize = NULL, folderColorRgb = NULL, fullFileExtension = NULL,
  hasAugmentedPermissions = NULL, hasThumbnail = NULL,
  headRevisionId = NULL, iconLink = NULL, id = NULL,
  imageMediaMetadata = NULL, indexableText = NULL, isAppAuthorized = NULL,
  labels = NULL, lastModifyingUser = NULL, lastModifyingUserName = NULL,
  lastViewedByMeDate = NULL, markedViewedByMeDate = NULL,
  md5Checksum = NULL, mimeType = NULL, modifiedByMeDate = NULL,
  modifiedDate = NULL, openWithLinks = NULL, originalFilename = NULL,
  ownedByMe = NULL, ownerNames = NULL, owners = NULL, parents = NULL,
  permissions = NULL, properties = NULL, quotaBytesUsed = NULL,
  selfLink = NULL, shareable = NULL, shared = NULL,
  sharedWithMeDate = NULL, sharingUser = NULL, spaces = NULL,
  teamDriveId = NULL, thumbnail = NULL, thumbnailLink = NULL,
  thumbnailVersion = NULL, title = NULL, trashedDate = NULL,
  trashingUser = NULL, userPermission = NULL, version = NULL,
  videoMediaMetadata = NULL, webContentLink = NULL, webViewLink = NULL,
  writersCanShare = NULL)
}
\arguments{
\item{File.capabilities}{The \link{File.capabilities} object or list of objects}

\item{File.exportLinks}{The \link{File.exportLinks} object or list of objects}

\item{File.imageMediaMetadata}{The \link{File.imageMediaMetadata} object or list of objects}

\item{File.imageMediaMetadata.location}{The \link{File.imageMediaMetadata.location} object or list of objects}

\item{File.indexableText}{The \link{File.indexableText} object or list of objects}

\item{File.labels}{The \link{File.labels} object or list of objects}

\item{File.openWithLinks}{The \link{File.openWithLinks} object or list of objects}

\item{File.thumbnail}{The \link{File.thumbnail} object or list of objects}

\item{File.videoMediaMetadata}{The \link{File.videoMediaMetadata} object or list of objects}

\item{alternateLink}{A link for opening the file in a relevant Google editor or viewer}

\item{appDataContents}{Whether this file is in the Application Data folder}

\item{canComment}{Whether the current user can comment on the file}

\item{canReadRevisions}{Whether the current user has read access to the Revisions resource of the file}

\item{capabilities}{Capabilities the current user has on the file}

\item{copyable}{Whether the file can be copied by the current user}

\item{createdDate}{Create time for this file (formatted RFC 3339 timestamp)}

\item{defaultOpenWithLink}{A link to open this file with the user's default app for this file}

\item{description}{A short description of the file}

\item{downloadUrl}{No description}

\item{editable}{Whether the file can be edited by the current user}

\item{embedLink}{A link for embedding the file}

\item{etag}{ETag of the file}

\item{explicitlyTrashed}{Whether this file has been explicitly trashed, as opposed to recursively trashed}

\item{exportLinks}{Links for exporting Google Docs to specific formats}

\item{fileExtension}{The final component of fullFileExtension with trailing text that does not appear to be part of the extension removed}

\item{fileSize}{The size of the file in bytes}

\item{folderColorRgb}{Folder color as an RGB hex string if the file is a folder}

\item{fullFileExtension}{The full file extension; extracted from the title}

\item{hasAugmentedPermissions}{Whether any users are granted file access directly on this file}

\item{hasThumbnail}{Whether this file has a thumbnail}

\item{headRevisionId}{The ID of the file's head revision}

\item{iconLink}{A link to the file's icon}

\item{id}{The ID of the file}

\item{imageMediaMetadata}{Metadata about image media}

\item{indexableText}{Indexable text attributes for the file (can only be written)}

\item{isAppAuthorized}{Whether the file was created or opened by the requesting app}

\item{labels}{A group of labels for the file}

\item{lastModifyingUser}{The last user to modify this file}

\item{lastModifyingUserName}{Name of the last user to modify this file}

\item{lastViewedByMeDate}{Last time this file was viewed by the user (formatted RFC 3339 timestamp)}

\item{markedViewedByMeDate}{Deprecated}

\item{md5Checksum}{An MD5 checksum for the content of this file}

\item{mimeType}{The MIME type of the file}

\item{modifiedByMeDate}{Last time this file was modified by the user (formatted RFC 3339 timestamp)}

\item{modifiedDate}{Last time this file was modified by anyone (formatted RFC 3339 timestamp)}

\item{openWithLinks}{A map of the id of each of the user's apps to a link to open this file with that app}

\item{originalFilename}{The original filename of the uploaded content if available, or else the original value of the title field}

\item{ownedByMe}{Whether the file is owned by the current user}

\item{ownerNames}{Name(s) of the owner(s) of this file}

\item{owners}{The owner(s) of this file}

\item{parents}{Collection of parent folders which contain this file}

\item{permissions}{The list of permissions for users with access to this file}

\item{properties}{The list of properties}

\item{quotaBytesUsed}{The number of quota bytes used by this file}

\item{selfLink}{A link back to this file}

\item{shareable}{Whether the file's sharing settings can be modified by the current user}

\item{shared}{Whether the file has been shared}

\item{sharedWithMeDate}{Time at which this file was shared with the user (formatted RFC 3339 timestamp)}

\item{sharingUser}{User that shared the item with the current user, if available}

\item{spaces}{The list of spaces which contain the file}

\item{teamDriveId}{ID of the Team Drive the file resides in}

\item{thumbnail}{A thumbnail for the file}

\item{thumbnailLink}{A short-lived link to the file's thumbnail}

\item{thumbnailVersion}{The thumbnail version for use in thumbnail cache invalidation}

\item{title}{The title of this file}

\item{trashedDate}{The time that the item was trashed (formatted RFC 3339 timestamp)}

\item{trashingUser}{If the file has been explicitly trashed, the user who trashed it}

\item{userPermission}{The permissions for the authenticated user on this file}

\item{version}{A monotonically increasing version number for the file}

\item{videoMediaMetadata}{Metadata about video media}

\item{webContentLink}{A link for downloading the content of the file in a browser using cookie based authentication}

\item{webViewLink}{A link only available on public folders for viewing their static web assets (HTML, CSS, JS, etc) via Google Drive's Website Hosting}

\item{writersCanShare}{Whether writers can share the document with other users}
}
\value{
File object
}
\description{
File Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata for a file.
}
\seealso{
Other File functions: \code{\link{File.capabilities}},
  \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{files.copy}}, \code{\link{files.insert}},
  \code{\link{files.patch}}, \code{\link{files.update}}
}
