% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.thumbnail}
\alias{File.thumbnail}
\title{File.thumbnail Object}
\usage{
File.thumbnail(image = NULL, mimeType = NULL)
}
\arguments{
\item{image}{The URL-safe Base64 encoded bytes of the thumbnail image}

\item{mimeType}{The MIME type of the thumbnail}
}
\value{
File.thumbnail object
}
\description{
File.thumbnail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A thumbnail for the file. This will only be used if Drive cannot generate a standard thumbnail.
}
\seealso{
Other File functions: \code{\link{File.capabilities}},
  \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.insert}}, \code{\link{files.patch}},
  \code{\link{files.update}}
}
