% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.videoMediaMetadata}
\alias{File.videoMediaMetadata}
\title{File.videoMediaMetadata Object}
\usage{
File.videoMediaMetadata(durationMillis = NULL, height = NULL,
  width = NULL)
}
\arguments{
\item{durationMillis}{The duration of the video in milliseconds}

\item{height}{The height of the video in pixels}

\item{width}{The width of the video in pixels}
}
\value{
File.videoMediaMetadata object
}
\description{
File.videoMediaMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata about video media. This will only be present for video types.
}
\seealso{
Other File functions: \code{\link{File.capabilities}},
  \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}}, \code{\link{File}},
  \code{\link{files.copy}}, \code{\link{files.insert}},
  \code{\link{files.patch}}, \code{\link{files.update}}
}
