% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{FileList}
\alias{FileList}
\title{FileList Object}
\usage{
FileList(etag = NULL, items = NULL, nextLink = NULL,
  nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{etag}{The ETag of the list}

\item{items}{The actual list of files}

\item{nextLink}{A link to the next page of files}

\item{nextPageToken}{The page token for the next page of files}

\item{selfLink}{A link back to this list}
}
\value{
FileList object
}
\description{
FileList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of files.
}

