% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.emptyTrash}
\alias{files.emptyTrash}
\title{Permanently deletes all of the user's trashed files.}
\usage{
files.emptyTrash()
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details. 

#' @importFrom googleAuthR gar_api_generator
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
