% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.get}
\alias{files.get}
\title{Gets a file's metadata by ID.}
\usage{
files.get(fileId, acknowledgeAbuse = NULL, projection = NULL,
  revisionId = NULL, updateViewedDate = NULL)
}
\arguments{
\item{fileId}{The ID for the file in question}

\item{acknowledgeAbuse}{Whether the user is acknowledging the risk of downloading known malware or other abusive files}

\item{projection}{This parameter is deprecated and has no function}

\item{revisionId}{Specifies the Revision ID that should be downloaded}

\item{updateViewedDate}{Deprecated: Use files}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.metadata.readonly, https://www.googleapis.com/auth/drive.photos.readonly, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}

