% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.insert}
\alias{files.insert}
\title{Insert a new file.}
\usage{
files.insert(File, convert = NULL, ocr = NULL, ocrLanguage = NULL,
  pinned = NULL, timedTextLanguage = NULL, timedTextTrackName = NULL,
  useContentAsIndexableText = NULL, visibility = NULL)
}
\arguments{
\item{File}{The \link{File} object to pass to this method}

\item{convert}{Whether to convert this file to the corresponding Google Docs format}

\item{ocr}{Whether to attempt OCR on}

\item{ocrLanguage}{If ocr is true, hints at the language to use}

\item{pinned}{Whether to pin the head revision of the uploaded file}

\item{timedTextLanguage}{The language of the timed text}

\item{timedTextTrackName}{The timed text track name}

\item{useContentAsIndexableText}{Whether to use the content as indexable text}

\item{visibility}{The visibility of the new file}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.apps.readonly
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.apps.readonly, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other File functions: \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.patch}}, \code{\link{files.update}}
}

