% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.list}
\alias{files.list}
\title{Lists the user's files.}
\usage{
files.list(corpora = NULL, corpus = NULL, includeTeamDriveItems = NULL,
  maxResults = NULL, orderBy = NULL, pageToken = NULL,
  projection = NULL, q = NULL, spaces = NULL, supportsTeamDrives = NULL,
  teamDriveId = NULL)
}
\arguments{
\item{corpora}{Comma-separated list of bodies of items (files/documents) to which the query applies}

\item{corpus}{The body of items (files/documents) to which the query applies}

\item{includeTeamDriveItems}{Whether Team Drive items should be included in results}

\item{maxResults}{Maximum number of files to return}

\item{orderBy}{A comma-separated list of sort keys}

\item{pageToken}{Page token for files}

\item{projection}{This parameter is deprecated and has no function}

\item{q}{Query string for searching files}

\item{spaces}{A comma-separated list of spaces to query}

\item{supportsTeamDrives}{Whether the requesting application supports Team Drives}

\item{teamDriveId}{ID of Team Drive to search}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.apps.readonly
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.apps.readonly, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.metadata.readonly, https://www.googleapis.com/auth/drive.photos.readonly, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
