% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.untrash}
\alias{files.untrash}
\title{Restores a file from the trash.}
\usage{
files.untrash(fileId, supportsTeamDrives = NULL)
}
\arguments{
\item{fileId}{The ID of the file to untrash}

\item{supportsTeamDrives}{Whether the requesting application supports Team Drives}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.apps.readonly
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.apps.readonly, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
