% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.update}
\alias{files.update}
\title{Updates file metadata and/or content.}
\usage{
files.update(File, fileId, addParents = NULL, convert = NULL,
  modifiedDateBehavior = NULL, newRevision = NULL, ocr = NULL,
  ocrLanguage = NULL, pinned = NULL, removeParents = NULL,
  setModifiedDate = NULL, supportsTeamDrives = NULL,
  timedTextLanguage = NULL, timedTextTrackName = NULL,
  updateViewedDate = NULL, useContentAsIndexableText = NULL)
}
\arguments{
\item{File}{The \link{File} object to pass to this method}

\item{fileId}{The ID of the file to update}

\item{addParents}{Comma-separated list of parent IDs to add}

\item{convert}{This parameter is deprecated and has no function}

\item{modifiedDateBehavior}{Determines the behavior in which modifiedDate is updated}

\item{newRevision}{Whether a blob upload should create a new revision}

\item{ocr}{Whether to attempt OCR on}

\item{ocrLanguage}{If ocr is true, hints at the language to use}

\item{pinned}{Whether to pin the new revision}

\item{removeParents}{Comma-separated list of parent IDs to remove}

\item{setModifiedDate}{Whether to set the modified date with the supplied modified date}

\item{supportsTeamDrives}{Whether the requesting application supports Team Drives}

\item{timedTextLanguage}{The language of the timed text}

\item{timedTextTrackName}{The timed text track name}

\item{updateViewedDate}{Whether to update the view date after successfully updating the file}

\item{useContentAsIndexableText}{Whether to use the content as indexable text}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.apps.readonly
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.scripts
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.apps.readonly, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.scripts)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other File functions: \code{\link{File.capabilities}},
  \code{\link{File.exportLinks}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.indexableText}},
  \code{\link{File.labels}},
  \code{\link{File.openWithLinks}},
  \code{\link{File.thumbnail}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.insert}}, \code{\link{files.patch}}
}
