% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{ParentList}
\alias{ParentList}
\title{ParentList Object}
\usage{
ParentList(etag = NULL, items = NULL, selfLink = NULL)
}
\arguments{
\item{etag}{The ETag of the list}

\item{items}{The list of parents}

\item{selfLink}{A link back to this list}
}
\value{
ParentList object
}
\description{
ParentList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of a file's parents.
}
