% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{parents.insert}
\alias{parents.insert}
\title{Adds a parent folder for a file.}
\usage{
parents.insert(ParentReference, fileId, supportsTeamDrives = NULL)
}
\arguments{
\item{ParentReference}{The \link{ParentReference} object to pass to this method}

\item{fileId}{The ID of the file}

\item{supportsTeamDrives}{Whether the requesting application supports Team Drives}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other ParentReference functions: \code{\link{ParentReference}}
}
