% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Permission}
\alias{Permission}
\title{Permission Object}
\usage{
Permission(additionalRoles = NULL, authKey = NULL, domain = NULL,
  emailAddress = NULL, etag = NULL, expirationDate = NULL, id = NULL,
  name = NULL, photoLink = NULL, role = NULL, selfLink = NULL,
  type = NULL, value = NULL, withLink = NULL)
}
\arguments{
\item{additionalRoles}{Additional roles for this user}

\item{authKey}{The authkey parameter required for this permission}

\item{domain}{The domain name of the entity this permission refers to}

\item{emailAddress}{The email address of the user or group this permission refers to}

\item{etag}{The ETag of the permission}

\item{expirationDate}{The time at which this permission will expire (RFC 3339 date-time)}

\item{id}{The ID of the user this permission refers to, and identical to the permissionId in the About and Files resources}

\item{name}{The name for this permission}

\item{photoLink}{A link to the profile photo, if available}

\item{role}{The primary role for this user}

\item{selfLink}{A link back to this permission}

\item{type}{The account type}

\item{value}{The email address or domain name for the entity}

\item{withLink}{Whether the link is required for this permission}
}
\value{
Permission object
}
\description{
Permission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A permission for a file.
}
\seealso{
Other Permission functions: \code{\link{permissions.insert}},
  \code{\link{permissions.patch}},
  \code{\link{permissions.update}}
}

