% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{PermissionId}
\alias{PermissionId}
\title{PermissionId Object}
\usage{
PermissionId(id = NULL)
}
\arguments{
\item{id}{The permission ID}
}
\value{
PermissionId object
}
\description{
PermissionId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An ID for a user or group as seen in Permission items.
}
