% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{PermissionList}
\alias{PermissionList}
\title{PermissionList Object}
\usage{
PermissionList(etag = NULL, items = NULL, selfLink = NULL)
}
\arguments{
\item{etag}{The ETag of the list}

\item{items}{The actual list of permissions}

\item{selfLink}{A link back to this list}
}
\value{
PermissionList object
}
\description{
PermissionList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of permissions associated with a file.
}

