% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{permissions.insert}
\alias{permissions.insert}
\title{Inserts a permission for a file.}
\usage{
permissions.insert(Permission, fileId, emailMessage = NULL,
  sendNotificationEmails = NULL)
}
\arguments{
\item{Permission}{The \link{Permission} object to pass to this method}

\item{fileId}{The ID for the file}

\item{emailMessage}{A custom message to include in notification emails}

\item{sendNotificationEmails}{Whether to send notification emails when sharing to users or groups}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other Permission functions: \code{\link{Permission}},
  \code{\link{permissions.patch}},
  \code{\link{permissions.update}}
}

