% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{properties.insert}
\alias{properties.insert}
\title{Adds a property to a file, or updates it if it already exists.}
\usage{
properties.insert(Property, fileId)
}
\arguments{
\item{Property}{The \link{Property} object to pass to this method}

\item{fileId}{The ID of the file}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other Property functions: \code{\link{Property}},
  \code{\link{properties.patch}},
  \code{\link{properties.update}}
}

