% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Property}
\alias{Property}
\title{Property Object}
\usage{
Property(etag = NULL, key = NULL, selfLink = NULL, value = NULL,
  visibility = NULL)
}
\arguments{
\item{etag}{ETag of the property}

\item{key}{The key of this property}

\item{selfLink}{The link back to this property}

\item{value}{The value of this property}

\item{visibility}{The visibility of this property}
}
\value{
Property object
}
\description{
Property Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A key-value pair attached to a file that is either public or private to an application.The following limits apply to file properties:  - Maximum of 100 properties total per file- Maximum of 30 private properties per app- Maximum of 30 public properties- Maximum of 124 bytes size limit on (key + value) string in UTF-8 encoding for a single property.
}
\seealso{
Other Property functions: \code{\link{properties.insert}},
  \code{\link{properties.patch}},
  \code{\link{properties.update}}
}
