% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{PropertyList}
\alias{PropertyList}
\title{PropertyList Object}
\usage{
PropertyList(etag = NULL, items = NULL, selfLink = NULL)
}
\arguments{
\item{etag}{The ETag of the list}

\item{items}{The list of properties}

\item{selfLink}{The link back to this list}
}
\value{
PropertyList object
}
\description{
PropertyList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of properties, key-value pairs that are either public or private to an application.
}
