% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{realtime.update}
\alias{realtime.update}
\title{Overwrites the Realtime API data model associated with this file with the provided JSON data model.}
\usage{
realtime.update(fileId, baseRevision = NULL)
}
\arguments{
\item{fileId}{The ID of the file that the Realtime API data model is associated with}

\item{baseRevision}{The revision of the model to diff the uploaded model against}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
