% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{replies.update}
\alias{replies.update}
\title{Updates an existing reply.}
\usage{
replies.update(CommentReply, fileId, commentId, replyId)
}
\arguments{
\item{CommentReply}{The \link{CommentReply} object to pass to this method}

\item{fileId}{The ID of the file}

\item{commentId}{The ID of the comment}

\item{replyId}{The ID of the reply}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other CommentReply functions: \code{\link{CommentReply}},
  \code{\link{replies.insert}}, \code{\link{replies.patch}}
}
