% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Revision.exportLinks}
\alias{Revision.exportLinks}
\title{Revision.exportLinks Object}
\usage{
Revision.exportLinks()
}
\value{
Revision.exportLinks object
}
\description{
Revision.exportLinks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Links for exporting Google Docs to specific formats.
}
\seealso{
Other Revision functions: \code{\link{Revision}},
  \code{\link{revisions.patch}},
  \code{\link{revisions.update}}
}
