% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{RevisionList}
\alias{RevisionList}
\title{RevisionList Object}
\usage{
RevisionList(etag = NULL, items = NULL, selfLink = NULL)
}
\arguments{
\item{etag}{The ETag of the list}

\item{items}{The actual list of revisions}

\item{selfLink}{A link back to this list}
}
\value{
RevisionList object
}
\description{
RevisionList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of revisions of a file.
}

