% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{StartPageToken}
\alias{StartPageToken}
\title{StartPageToken Object}
\usage{
StartPageToken(startPageToken = NULL)
}
\arguments{
\item{startPageToken}{The starting page token for listing changes}
}
\value{
StartPageToken object
}
\description{
StartPageToken Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
