% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{team.delete}
\alias{team.delete}
\title{Permanently deletes a Team Drive for which the user is an organizer. The Team Drive cannot contain any untrashed items.}
\usage{
team.delete(teamDriveId)
}
\arguments{
\item{teamDriveId}{The ID of the Team Drive}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
