% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{team.get}
\alias{team.get}
\title{Gets a Team Drive's metadata by ID.}
\usage{
team.get(teamDriveId)
}
\arguments{
\item{teamDriveId}{The ID of the Team Drive}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
