% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{User.picture}
\alias{User.picture}
\title{User.picture Object}
\usage{
User.picture(url = NULL)
}
\arguments{
\item{url}{A URL that points to a profile picture of this user}
}
\value{
User.picture object
}
\description{
User.picture Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The user's profile picture.
}
\seealso{
Other User functions: \code{\link{User}}
}
