% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{User}
\alias{User}
\title{User Object}
\usage{
User(User.picture = NULL, displayName = NULL, emailAddress = NULL,
  isAuthenticatedUser = NULL, permissionId = NULL, picture = NULL)
}
\arguments{
\item{User.picture}{The \link{User.picture} object or list of objects}

\item{displayName}{A plain text displayable name for this user}

\item{emailAddress}{The email address of the user}

\item{isAuthenticatedUser}{Whether this user is the same as the authenticated user for whom the request was made}

\item{permissionId}{The user's ID as visible in the permissions collection}

\item{picture}{The user's profile picture}
}
\value{
User object
}
\description{
User Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a Drive user.
}
\seealso{
Other User functions: \code{\link{User.picture}}
}

