% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Change}
\alias{Change}
\title{Change Object}
\usage{
Change(deleted = NULL, file = NULL, fileId = NULL, id = NULL,
  modificationDate = NULL, selfLink = NULL)
}
\arguments{
\item{deleted}{Whether the file has been deleted}

\item{file}{The updated state of the file}

\item{fileId}{The ID of the file associated with this change}

\item{id}{The ID of the change}

\item{modificationDate}{The time of this modification}

\item{selfLink}{A link back to this change}
}
\value{
Change object
}
\description{
Change Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Representation of a change to a file.
}

