% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{CommentReplyList}
\alias{CommentReplyList}
\title{CommentReplyList Object}
\usage{
CommentReplyList(items = NULL, nextLink = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{items}{The list of replies}

\item{nextLink}{A link to the next page of replies}

\item{nextPageToken}{The page token for the next page of replies}

\item{selfLink}{A link back to this list}
}
\value{
CommentReplyList object
}
\description{
CommentReplyList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of replies to a comment on a file in Google Drive.
}
